feat(plot_ntable): plot sample sizes by group instead of using a table
Thank you @brianlangseth-NOAA for having the idea of the original
table that spurred the thought that led to this figure and for
being willing to remove the table.

Eventually this code will be migrated into the nwfscSurvey package,
but for right now it is helpful here in lingcod.
Colors for the survey are color-blind friendly and could be a potential
source of colors for the fleet structure later on, but I am not worrying
about that for the data workshop.

The example is functional and produces the figure used for the
pre-assessment workshop.