Turn lingcod_2021 into a package, modeling CrumpLab/vertical
Separate package from analysis, using installed data that is processed
with scripts in the data-raw file.
Analyses will be in separate folders outside of the package structure
that are ignored in the build file.
Add more functions that will be helpful later on.
See README for overall structure but more commits are to come.