feat(unfit): a directory to store "unfit" work
Fit means being in shape enough to make the cut into a folder
or script touched by the R package.
Unfit doesn't mean it won't ever be fit, it just currently isn't there
so commit anything you want and if it proves fruitful later we can move
it to a folder that is accessed by the package.