update(catch_rec_CA_1980): filter MRFSS to remove first year
Email from John Budrick
The first year of the survey in 1980 I think they were still sorting
out sampling and expansion methods.
They are unreasonably high so some authors may use the average
of a few years that followed or other algorithms to supplant the values.
You might consider using what was there from the previous assessment and
split it up with Albin to match the new area boundaries.
That 1980 value is excessively high across assessments,
and it’s a systemic issue,
so being consistent with the last assessment seems reasonable.

Close #28