use sa4ss::add_figure rather than add_figure and use unique chunk names
2 fixes to changes from @brianlangseth-NOAA to be able to compile,
thank you Brian for using the style that was in the document to
add all of the text that you did. These two fixes were extremely
easy for me to do given the amount of time that you put in.
I am impressed that the document compiled right away after making
these. Thank you.