feat(compile_ling): one liner to build the document, it will probably break
Directory management for building the document. LOTS of things going on, but
it makes it where you don't have to change your working directory and you can
always stay in lingcod_2021 and not enter doc because the function does it for
you. It will also build the figures if you do not already have them.
Working for KFJ but I would like to test it with others who are game.
Get ready to be gamey though because this is likely to break. BETA VERSION!

See #101 for participation.