feat(ref_range): Fix bad figure reference when plot 3 of 3 doesn't exist for one area
CAAL for the WCGBTS has 3 plots in the South and only 2 plots in the North
this led to a bad figure reference for the range of figures. A function
now finds the figure labels for you. Use ref-range.
Needed to move the reading of the csv file for figures to the
beginning of the document so that it is searchable before the figures are
made.

Also use for discard and bodyweight figures