FIX!!!!!!!!!!! Major problem when submitting were these references.
Turns out the names are specific to area which is why they were not
working in the northern model but worked for the southern model.
string_switchletter will change any name with _[a-z]_ to the appropriate
letter for area given a word provided in newletter argument.
This works now for both areas.